// frontend/apps/pathways/app/profile/_modules/common/types.d.ts
export type User = {
  id: string;
  email: string;
  provider: string;
  platform: string;
  first_name: string;
  last_name: string;
  gender: string;
  ethnicity: string;
  nationality: string;
  phone_number: string;
  city: string;
  state: string;
  skills: string[];
  country: string;
  linkedin_url: string;
  instagram_url: string;
  facebook_url: string;
  x_twitter_url: string;
  personal_website_url: string;
  current_career: string;
  job_search_phase: string;
  goals: string[];
  interests: string[];
  career_summary: string;
  birthday: string;
  current_job_title: string;
  background_image_url: string;
  profile_picture_url?: string;
};